EDGEAI-359 changes to refactor grid rendering to shared file
