+++ src/js/combined.js
@@ -14,9 +14,12 @@ import Stats from "./Stats.js"
 import droppedframes from './droppedframes.js'
 import { dynamicSort } from './sort.js'
 import { parseNumbersInObject } from './parseNumbersInObject.js';
+import { OrbitControls } from './OrbitControls.js'
+import { clearThree, color_points_class, color_points_field } from './utils.js'
+import { grid_set_radarpoints, init_grid } from './grid_render.js'
+import { mask_colors } from './utils.js'
 
 const PI = Math.PI
-
 
 const stats = new Stats();
 const cameraPanel = stats.addPanel(new Stats.Panel('cameraFPS', '#fff', '#222'));
@@ -61,17 +64,7 @@ function fpsUpdate(panel, max) {
         lastUpdateTime = curr
     }
 }
-const mask_colors = [
-    new THREE.Color(1.0, 1.0, 1.0),
-    new THREE.Color(0., 1., 0.),
-    new THREE.Color(0.50980392, 0.50980392, 0.72941176),
-    new THREE.Color(0.00784314, 0.18823529, 0.29411765)
-]
 
-const mask_class_names = ["background",
-    "person",
-    "car", "truck"
-]
 
 const grid_scene = new THREE.Scene()
 grid_scene.background = new THREE.Color(0xa0a0a0)
@@ -189,12 +182,31 @@ function drawBoxesSpeedDistance(canvas, boxes, radar_points) {
                 ctx.fillText(lines[i], (box.center_x - box.width / 2) * canvas.width, (box.center_y - box.height / 2) * canvas.height + (lines.length - 1 - i * lineheight));
                 ctx.strokeText(lines[i], (box.center_x - box.width / 2) * canvas.width, (box.center_y - box.height / 2) * canvas.height + (lines.length - 1 - i * lineheight));
             }
-
         }
     }
 }
 
+const renderer_grid = new THREE.WebGLRenderer({ antialias: true, canvas: gridCanvas });
+let gridCanvasWidth = gridCanvas.parentElement.offsetWidth
+let gridCanvasHeight = gridCanvas.parentElement.offsetHeight
+renderer_grid.setSize(gridCanvasWidth, gridCanvasHeight)
+
+const HFOV = 82
+let aspect = gridCanvasWidth / gridCanvasHeight
+// let fov = Math.atan(Math.tan(HFOV * Math.PI / 360) / aspect) * 360 / Math.PI
+let fov = 20
+
+const camera_grid = new THREE.PerspectiveCamera(fov, aspect, 0.1, 1000);
+camera_grid.position.y = 1.9;
+camera_grid.position.z = -4;
+
+const orbitControls = new OrbitControls(camera_grid, gridCanvas);
+orbitControls.target = new THREE.Vector3(0, 0, 3.25);
+orbitControls.update();
+
 const loader = new THREE.FileLoader();
+
+let cameraNeedsUpdate = false;
 loader.load(
     // resource URL
     '/config/webui/details',
@@ -238,18 +250,11 @@ loader.load(
             socketUrlH264 = config.H264_TOPIC
         }
 
-        if (config.COMBINED_GRID_DRAW_PCD) {
-            GRID_DRAW_PCD = config.COMBINED_GRID_DRAW_PCD
-        }
-
         if (config.COMBINED_CAMERA_DRAW_PCD) {
             CAMERA_DRAW_PCD = config.COMBINED_CAMERA_DRAW_PCD
         }
         if (config.COMBINED_CAMERA_PCD_LABEL) {
             CAMERA_PCD_LABEL = config.COMBINED_CAMERA_PCD_LABEL
-        }
-        if (typeof config.DRAW_UNKNOWN_CELLS == "boolean") {
-            DRAW_UNKNOWN_CELLS = config.DRAW_UNKNOWN_CELLS
         }
 
         if (typeof config.DRAW_BOX == "boolean") {
@@ -260,10 +265,13 @@ loader.load(
             DRAW_BOX_TEXT = config.DRAW_BOX_TEXT
         }
 
+        config.GRID_DRAW_PCD = config.COMBINED_GRID_DRAW_PCD
+        init_grid(grid_scene, renderer_grid, camera_grid, config)
+
         const quad = new THREE.PlaneGeometry(width / height * 500, 500);
         const cameraUpdate = fpsUpdate(cameraPanel)
         h264Stream(socketUrlH264, 1920, 1080, 30, (timing) => {
-            cameraUpdate(), resetTimeout()
+            cameraUpdate(); resetTimeout(); cameraNeedsUpdate = true;
             // cameraMSPanel.update(timing.decode_time, 33) 
         }).then((tex) => {
             texture_camera = tex;
@@ -280,44 +288,7 @@ loader.load(
             scene.add(mesh_cam);
         })
 
-        alloc_bins()
-        const gridHelper = new PolarGridFan(RANGE_BIN_LIMITS[0], RANGE_BIN_LIMITS[1],
-            -ANGLE_BIN_LIMITS[0] * Math.PI / 180, -ANGLE_BIN_LIMITS[1] * Math.PI / 180,
-            Math.ceil((ANGLE_BIN_LIMITS[1] - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH),
-            Math.ceil((RANGE_BIN_LIMITS[1] - RANGE_BIN_LIMITS[0]) / RANGE_BIN_WIDTH),
-            64,
-            0x000,
-            0x000);
-        gridHelper.position.z = 0.002;
-        grid_scene.add(gridHelper);
-        let decimals = 1
-        if (Number.isInteger(RANGE_BIN_LIMITS[0]) && Number.isInteger(RANGE_BIN_WIDTH * 2)) {
-            decimals = 0
-        }
-        for (let i = RANGE_BIN_LIMITS[0]; i <= RANGE_BIN_LIMITS[1]; i += RANGE_BIN_WIDTH * 2) {
-            const myText = new SpriteText(i.toFixed(decimals) + "m", 0.03, "0x888888")           
-            myText.material.sizeAttenuation = false
-            myText.position.x = Math.sin((-ANGLE_BIN_LIMITS[0]+1) / 180 * PI) * i + Math.sin((-ANGLE_BIN_LIMITS[0] + 91) / 180 * PI)*0.16
-            myText.position.z = Math.cos((-ANGLE_BIN_LIMITS[0]+1) / 180 * PI) * i + Math.cos((-ANGLE_BIN_LIMITS[0] + 91) / 180 * PI)*0.16
-            grid_scene.add(myText)
-        }
-        for (let i = RANGE_BIN_LIMITS[0]; i <= RANGE_BIN_LIMITS[1]; i += RANGE_BIN_WIDTH * 2) {
-            const myText = new SpriteText(i.toFixed(decimals) + "m", 0.03, "0x888888")
-            myText.material.sizeAttenuation = false
-            myText.position.x = Math.sin((-ANGLE_BIN_LIMITS[1]-1) / 180 * PI) * i + Math.sin((-ANGLE_BIN_LIMITS[1] - 91) / 180 * PI) * 0.16
-            myText.position.z = Math.cos((-ANGLE_BIN_LIMITS[1]-1) / 180 * PI) * i + Math.cos((-ANGLE_BIN_LIMITS[1] - 91) / 180 * PI) * 0.16
-            grid_scene.add(myText)
-        }
-
-        for (let i = ANGLE_BIN_LIMITS[0]; i <= ANGLE_BIN_LIMITS[1]; i += ANGLE_BIN_WIDTH * 2) {
-            let pad = i < 0 ? "" : " "
-            const myText = new SpriteText(pad + i.toFixed(0) + "°", 0.03, "0x888888")
-            myText.material.sizeAttenuation = false
-            myText.position.x = Math.sin(-i / 180 * PI) * RANGE_BIN_LIMITS[1]
-            myText.position.y = 0.2
-            myText.position.z = Math.cos(-i / 180 * PI) * RANGE_BIN_LIMITS[1]
-            grid_scene.add(myText)
-        }
+        
 
         const modelFPSUpdate = fpsUpdate(modelPanel)
 
@@ -364,6 +335,7 @@ loader.load(
             radar_points.points = filteredPoints
         }).then((pcd) => {
             radar_points = pcd;
+            grid_set_radarpoints(radar_points)
         })
     },
     function (xhr) {
@@ -375,148 +347,10 @@ loader.load(
 );
 THREE.Cache.enabled = true;
 
-const bins = []
-var window_index = 0
-function alloc_bins() {
-    bins.length = 0
-    const RANGE_BIN_COUNT = RANGE_BIN_LIMITS[1] / RANGE_BIN_WIDTH - RANGE_BIN_LIMITS[0] / RANGE_BIN_WIDTH + 1
-    for (let i = ANGLE_BIN_LIMITS[0]; i <= ANGLE_BIN_LIMITS[1]; i += ANGLE_BIN_WIDTH) {
-        let n = Array(RANGE_BIN_COUNT)
-        for (let j = 0; j < RANGE_BIN_COUNT; j++) {
-            n[j] = Array(WINDOW_LENGTH).fill([])
-        }
-        bins.push(n)
-    }
-    return bins
-}
 
 
-function clear_bins() {
-    window_index = (window_index + 1) % WINDOW_LENGTH
-    for (let i = 0; i < bins.length; i++) {
-        for (let j = 0; j < bins[i].length; j++) {
-            bins[i][j][window_index] = []
-        }
-    }
-}
 
-function get_bin(angle, range) {
-    if (angle < ANGLE_BIN_LIMITS[0]) {
-        angle = ANGLE_BIN_LIMITS[0]
-    }
-    if (angle > ANGLE_BIN_LIMITS[1]) {
-        angle = ANGLE_BIN_LIMITS[1]
-    }
-    if (range < RANGE_BIN_LIMITS[0]) {
-        range = RANGE_BIN_LIMITS[0]
-    }
-    if (range > RANGE_BIN_LIMITS[1]) {
-        range = RANGE_BIN_LIMITS[1]
-    }
-    const i = Math.floor((angle - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH)
-    const j = Math.floor((range - RANGE_BIN_LIMITS[0]) / RANGE_BIN_WIDTH)
-    return bins[i][j][window_index]
-
-}
-
-function increment_bin(angle, range, value) {
-    if (angle < ANGLE_BIN_LIMITS[0]) {
-        angle = ANGLE_BIN_LIMITS[0]
-    }
-    if (angle > ANGLE_BIN_LIMITS[1]) {
-        angle = ANGLE_BIN_LIMITS[1]
-    }
-    if (range < RANGE_BIN_LIMITS[0]) {
-        range = RANGE_BIN_LIMITS[0]
-    }
-    if (range > RANGE_BIN_LIMITS[1]) {
-        range = RANGE_BIN_LIMITS[1]
-    }
-    const i = Math.floor((angle - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH)
-    const j = Math.floor((range - RANGE_BIN_LIMITS[0]) / RANGE_BIN_WIDTH)
-    bins[i][j][window_index].push(value)
-}
-
-function getValsInBin(angle, range, angleBinOffset, rangeBinOffset) {
-    let angleBin = Math.floor((angle - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH) + angleBinOffset
-    let rangeBin = Math.floor((range - RANGE_BIN_LIMITS[0]) / RANGE_BIN_WIDTH) + rangeBinOffset
-    angleBin = (angleBin + bins.length) % bins.length
-    if (rangeBin < 0) {
-        // rangeBin = 0
-        return []
-    }
-    if (rangeBin >= bins[angleBin].length) {
-        // rangeBin = bins[angleBin].length - 1
-        return []
-    }
-    let sum = bins[angleBin][rangeBin].reduce((a, b) => a.concat(b), [])
-    return sum
-}
-
-function getCountInBin(angle, range, angleBinOffset, rangeBinOffset) {
-    let angleBin = Math.floor((angle - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH) + angleBinOffset
-    let rangeBin = Math.floor((range - RANGE_BIN_LIMITS[0]) / RANGE_BIN_WIDTH) + rangeBinOffset
-    angleBin = (angleBin + bins.length) % bins.length
-    if (rangeBin < 0) {
-        // rangeBin = 0
-        return 0
-    }
-    if (rangeBin >= bins[angleBin].length) {
-        // rangeBin = bins[angleBin].length - 1
-        return 0
-    }
-    let sum = bins[angleBin][rangeBin].reduce((a, b) => a + b.length, 0)
-    return sum
-}
-
-function getClassInList(l) {
-    const classes = {}
-    l.forEach((point) => {
-        if (isFinite(classes[point.class])) {
-            classes[point.class] += 1
-        } else {
-            classes[point.class] = 1
-        }
-    })
-    if (classes[0] == l.length) {
-        return [0, classes[0]]
-    }
-    classes[0] = 0
-    
-    let max_class = 0
-    let max_class_val = -1
-    for (var cl in classes) {
-        if (max_class_val < classes[cl]) {
-            max_class = cl
-            max_class_val = classes[cl]
-        }
-    }
-    return [max_class, max_class_val]
-}
-
-
-function clearThree(obj) {
-    while (obj.children.length > 0) {
-        clearThree(obj.children[0]);
-    }
-    if (obj.geometry) obj.geometry.dispose();
-
-    if (obj.material) {
-        //in case of map, bumpMap, normalMap, envMap ...
-        Object.keys(obj.material).forEach(prop => {
-            if (!obj.material[prop])
-                return;
-            if (obj.material[prop] !== null && typeof obj.material[prop].dispose === 'function')
-                obj.material[prop].dispose();
-        })
-        obj.material.dispose();
-    }
-    obj.parent.remove(obj);
-    obj.removeFromParent()
-}
-
-
-const rendered_boxes = []
+const rendered = []
 
 renderer.setAnimationLoop(animate);
 
@@ -524,38 +358,20 @@ renderer.setAnimationLoop(animate);
 function animate() {
     // animationUpdate()
 
-    if (typeof detect_boxes !== "undefined" && detect_boxes.needsUpdate) {
-        rendered_boxes.forEach((cell) => {
-            clearThree(cell)
-        })
-        rendered_boxes.length = 1
-        detect_boxes.msg.boxes.forEach((box) => {
-            // console.log(box)
-            const scale = 10
-            const pos = []
-            pos.push(box.width * scale * width / height / 2, box.height * scale / 2, 0);
-            pos.push(-box.width * scale * width / height / 2, box.height * scale / 2, 0);
-            pos.push(-box.width * scale * width / height / 2, -box.height * scale / 2, 0);
-            pos.push(box.width * scale * width / height / 2, -box.height * scale / 2, 0);
-            pos.push(box.width * scale * width / height / 2, box.height * scale / 2, 0);
-            const geometry = new LineGeometry();
-            geometry.setPositions(pos);
-            const rendered_box = new Line2(geometry, new LineMaterial({ color: 0x00ff00, linewidth: 10 }));
-            rendered_box.rotation.x = PI / 2;
-            rendered_box.position.x = (box.center_x - 0.5) * scale * width / height
-            rendered_box.position.z = (0.5 - box.center_y) * scale
-            rendered_box.position.y = scale / Math.tan(46.4 / 2 / 180 * PI) / 2
-            rendered_boxes.push(rendered_box)
-            scene.add(rendered_box)
-        })
+    if ((typeof mask_tex !== "undefined" || typeof detect_boxes !== "undefined") && typeof radar_points !== "undefined") {
+        if (CAMERA_DRAW_PCD != "disabled" && radar_points.points.length > 0) {
+            let points = radar_points.points
+            rendered.forEach((cell) => {
+                clearThree(cell)
+            })
+            if (CAMERA_DRAW_PCD == "class") {
+                color_points_class(points, scene, rendered, true, CAMERA_PCD_LABEL)
+            } else {
+                color_points_field(points, CAMERA_DRAW_PCD, scene, rendered, true, CAMERA_PCD_LABEL)
+            }
+        }
     }
-
-    // camera.rotation.y += 0.02
-    if (texture_camera) {
-        texture_camera.needsUpdate = true;
-    }
-
-    renderer.render(scene, camera);
+    renderer.render(scene, camera)
 
 }
 
@@ -568,268 +384,6 @@ function resetTimeout() {
     }, 15000);
 }
 
-
-
-const renderer_grid = new THREE.WebGLRenderer({ antialias: true, canvas: gridCanvas });
-// canvas.width = canvas.offsetWidth;
-// canvas.height = canvas.offsetHeight;
-// renderer_grid.setSize(540, 1080)
-// renderer.domElement.style.cssText = ""
-let gridCanvasWidth = gridCanvas.parentElement.offsetWidth
-let gridCanvasHeight = gridCanvas.parentElement.offsetHeight
-renderer_grid.setSize(gridCanvasWidth, gridCanvasHeight)
-function newRingGeo(angle, range, class_) {
-    const geometry = new THREE.RingGeometry(range, range + RANGE_BIN_WIDTH, 4, 1, angle / 180 * PI + PI / 2, ANGLE_BIN_WIDTH / 180 * PI);
-    const color = mask_colors[class_]
-    const material = new THREE.MeshBasicMaterial({ color: color, side: THREE.DoubleSide });
-    const mesh = new THREE.Mesh(geometry, material);
-    mesh.rotation.x = PI / 2
-    // mesh.rotation.z = PI / 2
-    return mesh
-}
-import { OrbitControls } from './OrbitControls.js'
-import { PolarGridFan } from './polarGridFan.js'
-
-
-const HFOV = 82
-let aspect = gridCanvasWidth / gridCanvasHeight
-// let fov = Math.atan(Math.tan(HFOV * Math.PI / 360) / aspect) * 360 / Math.PI
-let fov = 20
-
-const camera_grid = new THREE.PerspectiveCamera(fov, aspect, 0.1, 1000);
-camera_grid.position.y = 1.9;
-camera_grid.position.z = -4;
-
-const orbitControls = new OrbitControls(camera_grid, gridCanvas);
-orbitControls.target = new THREE.Vector3(0, 0, 3.25);
-orbitControls.update();
-
-renderer_grid.setAnimationLoop(animate_grid);
-
-const occupied = []
-const rendered_points = []
-
-function color_points_field(points, field, scene, height = false, label = "disabled") {
-    points.sort(dynamicSort(field))
-    let min_val = points[0][field]
-    let max_val = points[points.length - 1][field]
-    let avg_val = points[Math.floor(points.length / 2)][field]
-
-    let maxDelta = Math.max(avg_val - min_val, max_val - min_val)
-    min_val = avg_val - maxDelta
-    max_val = avg_val + maxDelta
-    points.forEach((point) => {
-        const geometry = new THREE.SphereGeometry(0.1)
-        let color = new THREE.Color(0xFFFFFF)
-        if (max_val - min_val > 0) {
-            if (point[field] < avg_val) {
-                let l = (point[field] - avg_val) / (min_val - avg_val)
-                l = (2 - l) / 2 * 100
-                color = new THREE.Color("hsl(240, 100%, " + l + "%)")
-            } else if (point[field] > avg_val) {
-                let l = (point[field] - avg_val) / (max_val - avg_val)
-                l = (2 - l) / 2 * 100
-                color = new THREE.Color("hsl(0, 100%, " + l + "%)")
-            }
-        }
-        const material = new THREE.MeshBasicMaterial({ color });
-        const sphere = new THREE.Mesh(geometry, material);
-        sphere.position.x = point.y
-        sphere.position.z = point.x
-        if (height) {
-            sphere.position.y = point.z
-        }
-        rendered_points.push(sphere)
-        scene.add(sphere)
-
-        if (label != "disabled") {
-            const myText = new SpriteText(point[label].toFixed(2), 0.025, "0x888888")
-            myText.material.sizeAttenuation = false
-            const factor = 1 - 0.12 / Math.sqrt(point.y * point.y + point.x * point.x)
-            myText.position.x = point.y * factor
-            myText.position.z = point.x * factor
-            if (height) {
-                myText.position.y = point.z
-            }
-            myText.position.y += 0.12
-            scene.add(myText)
-            rendered_points.push(myText)
-        }
-    })
-}
-
-function color_points_class(points, scene, height = false, label = "disabled") {
-    points.forEach((point) => {
-        const geometry = new THREE.SphereGeometry(0.1)
-        let color = new THREE.Color(0xFFFFFF)
-        if (point.class > 0) {
-            color = mask_colors[point.class]
-        }
-        const material = new THREE.MeshBasicMaterial({ color });
-        const sphere = new THREE.Mesh(geometry, material);
-        sphere.position.x = point.y
-        sphere.position.z = point.x
-        if (height) {
-            sphere.position.y = point.z
-        }
-        rendered_points.push(sphere)
-        scene.add(sphere)
-
-        if (label != "disabled") {
-            const myText = new SpriteText(point[label], 0.025, "0x888888")
-            myText.material.sizeAttenuation = false
-            const factor = 1 - 0.12 / Math.sqrt(point.y * point.y + point.x * point.x)
-            myText.position.x = point.y * factor
-            myText.position.z = point.x * factor
-            if (height) {
-                myText.position.y = point.z
-            }
-            myText.position.y += 0.12
-            scene.add(myText)
-            rendered_points.push(myText)
-        }
-    })
-}
-
-function animate_grid() {
-    if ((typeof mask_tex !== "undefined" || typeof detect_boxes !== "undefined") && typeof radar_points !== "undefined" && radar_points.needsUpdate && bins.length > 0) {
-        clear_bins()
-        rendered_points.forEach((cell) => {
-            clearThree(cell)
-            grid_scene.remove(cell)
-            scene.remove(cell)
-        })
-        rendered_points.length = 0
-        let points = radar_points.points
-        if (GRID_DRAW_PCD != "disabled" && radar_points.points.length > 0) {
-            if (GRID_DRAW_PCD == "class") {
-                color_points_class(points, grid_scene)
-            } else {
-                color_points_field(points, GRID_DRAW_PCD, grid_scene)
-            }
-        }
-
-        if (CAMERA_DRAW_PCD != "disabled" && radar_points.points.length > 0) {
-            if (CAMERA_DRAW_PCD == "class") {
-                color_points_class(points, scene, true, CAMERA_PCD_LABEL)
-            } else {
-                color_points_field(points, CAMERA_DRAW_PCD, scene, true, CAMERA_PCD_LABEL)
-            }
-        }
-
-        for (let p of points) {
-            if (DRAW_UNKNOWN_CELLS || p.class > 0) {
-                p.range = Math.sqrt(p.x * p.x + p.y * p.y + p.z * p.z)
-                p.angle = Math.atan2(p.y, p.x)
-                increment_bin(-p.angle * 180 / PI, p.range, p)
-            }
-        }
-        occupied.forEach((cell) => {
-            clearThree(cell)
-            grid_scene.remove(cell)
-        })
-        occupied.length = 0
-        let foundOccupied = new Array(bins.length).fill(false)
-        for (let j = RANGE_BIN_LIMITS[0]; j <= RANGE_BIN_LIMITS[1]; j += RANGE_BIN_WIDTH) {
-            for (let i = ANGLE_BIN_LIMITS[0] + ANGLE_BIN_WIDTH; i <= ANGLE_BIN_LIMITS[1] - ANGLE_BIN_WIDTH; i += ANGLE_BIN_WIDTH) {
-                let currInd = (i - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH
-                if (foundOccupied[currInd]) {
-                    continue;
-                }
-                let val0 = getValsInBin(i, j, 0, 0)
-                let val1 = getValsInBin(i, j, 0, -1)
-                let val2 = getValsInBin(i, j, 0, -2)
-                let sum0 = val0.length
-                let sum1 = val1.length
-                let sum2 = val2.length
-                if (sum0 >= BIN_THRESHOLD) {
-                    const [class_, _] = getClassInList(val0)
-                    const cell = newRingGeo(i, j, class_)
-                    occupied.push(cell)
-                    grid_scene.add(cell)
-                    foundOccupied[currInd] = true
-                    continue
-                }
-                if (sum0 + sum1 >= BIN_THRESHOLD) {
-                    const [class_, _] = getClassInList(val0.concat(val1))
-                    const cell = newRingGeo(i, j - RANGE_BIN_WIDTH, class_)
-                    occupied.push(cell)
-                    grid_scene.add(cell)
-                    foundOccupied[currInd] = true
-                    continue
-                }
-                if (sum0 + sum1 + sum2 >= BIN_THRESHOLD) {
-                    const [class_, _] = getClassInList(val0.concat(val1, val2))
-                    const cell = newRingGeo(i, j - RANGE_BIN_WIDTH * 2, class_)
-                    occupied.push(cell)
-                    grid_scene.add(cell)
-                    foundOccupied[currInd] = true
-                    continue
-                }
-            }
-        }
-
-        for (let j = RANGE_BIN_LIMITS[0]; j <= RANGE_BIN_LIMITS[1]; j += RANGE_BIN_WIDTH) {
-            for (let i = ANGLE_BIN_LIMITS[0] + ANGLE_BIN_WIDTH; i <= ANGLE_BIN_LIMITS[1] - ANGLE_BIN_WIDTH; i += ANGLE_BIN_WIDTH) {
-                let angle = i + ANGLE_BIN_WIDTH / 2
-                let currInd = (i - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH
-                if (foundOccupied[currInd]) {
-                    continue;
-                }
-                let val0 = getValsInBin(i, j, 0, 0)
-                let val1 = getValsInBin(i, j, 0, -1)
-                let val2 = getValsInBin(i, j, 0, -2)
-
-                if (!foundOccupied[(i - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH + 1]) {
-                    val0 = val0.concat(getValsInBin(i, j, 1, 0))
-                    val1 = val1.concat(getValsInBin(i, j, 1, -1))
-                    val2 = val2.concat(getValsInBin(i, j, 1, -2))
-                }
-                if (!foundOccupied[((i - ANGLE_BIN_LIMITS[0]) / ANGLE_BIN_WIDTH - 1 + bins.length) % bins.length]) {
-                    val0 = val0.concat(getValsInBin(i, j, -1, 0))
-                    val1 = val1.concat(getValsInBin(i, j, -1, -1))
-                    val2 = val2.concat(getValsInBin(i, j, -1, -2))
-                }
-                let sum0 = val0.length
-                let sum1 = val1.length
-                let sum2 = val2.length
-                if (sum0 >= BIN_THRESHOLD) {
-                    const [class_, _] = getClassInList(val0)
-                    const cell = newRingGeo(i, j, class_)
-                    occupied.push(cell)
-                    grid_scene.add(cell)
-                    foundOccupied[currInd] = true
-                    if (currInd > 0) { foundOccupied[currInd - 1] = true }
-                    if (currInd < bins.length - 1) { foundOccupied[currInd + 1] = true }
-                    continue
-                }
-                if (sum0 + sum1 >= BIN_THRESHOLD) {
-                    const [class_, _] = getClassInList(val0.concat(val1))
-                    const cell = newRingGeo(i, j - RANGE_BIN_WIDTH, class_)
-                    occupied.push(cell)
-                    grid_scene.add(cell)
-                    foundOccupied[currInd] = true
-                    if (currInd > 0) { foundOccupied[currInd - 1] = true }
-                    if (currInd < bins.length - 1) { foundOccupied[currInd + 1] = true }
-                    continue
-                }
-                if (sum0 + sum1 + sum2 >= BIN_THRESHOLD) {
-                    const [class_, _] = getClassInList(val0.concat(val1, val2))
-                    const cell = newRingGeo(i, j - RANGE_BIN_WIDTH * 2, class_)
-                    occupied.push(cell)
-                    grid_scene.add(cell)
-                    foundOccupied[currInd] = true
-                    if (currInd > 0) { foundOccupied[currInd - 1] = true }
-                    if (currInd < bins.length - 1) { foundOccupied[currInd + 1] = true }
-                    continue
-                }
-            }
-        }
-
-        radar_points.needsUpdate = false
-    }
-    renderer_grid.render(grid_scene, camera_grid);
-}
 
 window.addEventListener('resize', onWindowResize);
 function onWindowResize() {
